+++ src/RestApi/Device/Alive/DeviceAliveManagement.ts
@@ -3,7 +3,8 @@ import IOptions from '../../IOptions';
 import { getResource, parseJSONResponse } from '../../requester';
 import { Resources } from '../../resources';
 import IDevice from '../IDevice';
-import DeviceAlive, { IDeviceAlive } from './DeviceAlive';
+import DeviceAlive, { IDeviceAlive, IDeviceAliveList } from './DeviceAlive';
+import { Paginator } from '../../../Lib/Pagination/paginator';
 
 interface IDeviceAliveFilter {
 	deviceUids?: IDevice['uid'][];
@@ -18,17 +19,18 @@ interface IDeviceAliveGetParams {
 }
 
 export default class DeviceAliveManagement {
-	constructor(private options: IOptions) {}
+	constructor(
+		private options: IOptions,
+		private paginator: Paginator,
+	) {}
 
-	public async list({ filter, sort, pagination }: IDeviceAliveListParams) {
-		const devicesAlive = await getResource(this.options, `${Resources.Device}/alive`, {
+	public async list({ filter, sort, pagination }: IDeviceAliveListParams): Promise<IDeviceAliveList> {
+		const resp = await getResource(this.options, `${Resources.Device}/alive`, {
 			...(filter ?? {}),
 			...(sort ?? {}),
 			...(pagination ?? {}),
 		});
-		const devicesAliveParsed: IDeviceAlive[] = await parseJSONResponse(devicesAlive);
-
-		return devicesAliveParsed.map((deviceAlive) => new DeviceAlive(deviceAlive));
+		return this.paginator.getPaginatedListFromResponse(resp, (data) => new DeviceAlive(data));
 	}
 
 	public async get({ uid }: IDeviceAliveGetParams) {